Move some methods into api-call
