+++ src/channel-list-item.tsx
@@ -10,7 +10,7 @@ import { Action } from './lib/action';
 import { ChannelBase } from './lib/models/api-shapes';
 import { ChannelListViewModel } from './channel-list';
 import { fromObservable, Model } from './lib/model';
-import { isDM } from './channel-utils';
+import { isDM } from './lib/models/api-call';
 import { SimpleView } from './lib/view';
 import { Store } from './store';
 import { Updatable } from './lib/updatable';
