+++ src/channel-list.tsx
@@ -3,7 +3,7 @@ import * as React from 'react';
 
 import { ChannelBase } from './lib/models/api-shapes';
 import { ChannelViewModel, ChannelListItem } from './channel-list-item';
-import { channelSort, isDM } from './channel-utils';
+import { channelSort, isDM } from './lib/models/api-call';
 import { CollectionView } from './lib/collection-view';
 import { fromObservable, notify, Model } from './lib/model';
 import { Store, ChannelList } from './store';
